# rvgo: experimental no-alloc noescape merkle funcs