--- a/rvgo/fast/page.go
+++ b/rvgo/fast/page.go
@@ -3,8 +3,6 @@ package fast
 import (
 	"encoding/hex"
 	"fmt"
-
-	"github.com/ethereum/go-ethereum/crypto"
 )
 
 type Page [PageSize]byte
@@ -55,7 +53,7 @@ func (p *CachedPage) MerkleRoot() [32]byte {
 		if p.Ok[j] {
 			continue
 		}
-		p.Cache[j] = crypto.Keccak256Hash(p.Data[i : i+64])
+		p.Cache[j] = hash([64]byte(p.Data[i : i+64]))
 		//fmt.Printf("0x%x 0x%x -> 0x%x\n", p.Data[i:i+32], p.Data[i+32:i+64], p.Cache[j])
 		p.Ok[j] = true
 	}
@@ -66,7 +64,7 @@ func (p *CachedPage) MerkleRoot() [32]byte {
 		if p.Ok[j] {
 			continue
 		}
-		p.Cache[j] = HashPair(p.Cache[i], p.Cache[i+1])
+		p.Cache[j] = hashPair(p.Cache[i], p.Cache[i+1])
 		p.Ok[j] = true
 	}
 
@@ -100,11 +98,10 @@ func (p *CachedPage) MerkleizeNode(addr, gindex uint64) [32]byte {
 	return p.Cache[gindex]
 }
 
-func (p *CachedPage) GenerateProof(addr uint64) [][32]byte {
+func (p *CachedPage) GenerateProof(addr uint64, proofs *[8][32]byte) {
 	// Page-level proof
 	pageGindex := PageSize>>5 + (addr&PageAddrMask)>>5
 
-	proofs := make([][32]byte, 8)
 	proofIndex := 0
 
 	proofs[proofIndex] = p.MerkleizeSubtree(pageGindex)
@@ -114,6 +111,4 @@ func (p *CachedPage) GenerateProof(addr uint64) [][32]byte {
 		proofIndex++
 		proofs[proofIndex] = p.MerkleizeSubtree(uint64(sibling))
 	}
-
-	return proofs
 }