**Description**

A little bit hacky. And makes the assumption there's only 1 concurrent hasher (this could be fixed by allocating a `keccakState` per Memory merkleization job, and passing it as typed function argument through the merkle functions).

- Share hasher, so hasher doesn't get re-allocated every hash call
- Tricks with `go:noescape`  and `go:linkname` to force it to not assume the function interface, and not allow any slice data to escape to the heap)

This essentially removes all the non-essential allocations from the merkleization work, in a bit of an unsafe non-recommended way, to see how much performance can improve with this kind of heap-analysis informed memory optimization.

Profiles of the benchmark show that it's only a little bit faster, even after removing a theorectical ~10 million object allocations.

Before:
![before](https://github.com/user-attachments/assets/bbc9adcc-6e9b-4c13-bcd7-d6819ccffddc)

After:
![after](https://github.com/user-attachments/assets/848458fc-e36e-4beb-b0b9-c2ed5d76bb60)

CPU:
Before:
```
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkMemoryOperations
BenchmarkMemoryOperations/RandomReadWrite_Small
BenchmarkMemoryOperations/RandomReadWrite_Small-32         	48573948	        23.16 ns/op
BenchmarkMemoryOperations/RandomReadWrite_Medium
BenchmarkMemoryOperations/RandomReadWrite_Medium-32        	   12456	    126946 ns/op
BenchmarkMemoryOperations/RandomReadWrite_Large
BenchmarkMemoryOperations/RandomReadWrite_Large-32         	    9952	    112717 ns/op
BenchmarkMemoryOperations/SequentialReadWrite_Small
BenchmarkMemoryOperations/SequentialReadWrite_Small-32     	231982236	         5.219 ns/op
BenchmarkMemoryOperations/SequentialReadWrite_Large
BenchmarkMemoryOperations/SequentialReadWrite_Large-32     	222372898	         5.442 ns/op
BenchmarkMemoryOperations/SparseMemoryUsage
BenchmarkMemoryOperations/SparseMemoryUsage-32             	  528241	      2216 ns/op
BenchmarkMemoryOperations/DenseMemoryUsage
BenchmarkMemoryOperations/DenseMemoryUsage-32              	141923272	         8.803 ns/op
BenchmarkMemoryOperations/SmallFrequentUpdates
BenchmarkMemoryOperations/SmallFrequentUpdates-32          	36277586	        32.13 ns/op
BenchmarkMemoryOperations/MerkleProofGeneration_Small
BenchmarkMemoryOperations/MerkleProofGeneration_Small-32   	  655786	      1573 ns/op
BenchmarkMemoryOperations/MerkleProofGeneration_Large
BenchmarkMemoryOperations/MerkleProofGeneration_Large-32   	  561040	      1985 ns/op
BenchmarkMemoryOperations/MerkleRootCalculation_Small
BenchmarkMemoryOperations/MerkleRootCalculation_Small-32   	224840816	         5.409 ns/op
BenchmarkMemoryOperations/MerkleRootCalculation_Large
BenchmarkMemoryOperations/MerkleRootCalculation_Large-32   	205111311	         5.721 ns/op
PASS
```

After:
```
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkMemoryOperations
BenchmarkMemoryOperations/RandomReadWrite_Small
BenchmarkMemoryOperations/RandomReadWrite_Small-32         	51766090	        21.48 ns/op
BenchmarkMemoryOperations/RandomReadWrite_Medium
BenchmarkMemoryOperations/RandomReadWrite_Medium-32        	   12172	    130552 ns/op
BenchmarkMemoryOperations/RandomReadWrite_Large
BenchmarkMemoryOperations/RandomReadWrite_Large-32         	   10000	    113670 ns/op
BenchmarkMemoryOperations/SequentialReadWrite_Small
BenchmarkMemoryOperations/SequentialReadWrite_Small-32     	223583344	         4.971 ns/op
BenchmarkMemoryOperations/SequentialReadWrite_Large
BenchmarkMemoryOperations/SequentialReadWrite_Large-32     	232579297	         5.322 ns/op
BenchmarkMemoryOperations/SparseMemoryUsage
BenchmarkMemoryOperations/SparseMemoryUsage-32             	  423361	      2386 ns/op
BenchmarkMemoryOperations/DenseMemoryUsage
BenchmarkMemoryOperations/DenseMemoryUsage-32              	145170620	         8.592 ns/op
BenchmarkMemoryOperations/SmallFrequentUpdates
BenchmarkMemoryOperations/SmallFrequentUpdates-32          	36379844	        30.87 ns/op
BenchmarkMemoryOperations/MerkleProofGeneration_Small
BenchmarkMemoryOperations/MerkleProofGeneration_Small-32   	  702778	      1490 ns/op
BenchmarkMemoryOperations/MerkleProofGeneration_Large
BenchmarkMemoryOperations/MerkleProofGeneration_Large-32   	  555963	      1900 ns/op
BenchmarkMemoryOperations/MerkleRootCalculation_Small
BenchmarkMemoryOperations/MerkleRootCalculation_Small-32   	233706420	         5.124 ns/op
BenchmarkMemoryOperations/MerkleRootCalculation_Large
BenchmarkMemoryOperations/MerkleRootCalculation_Large-32   	218443596	         5.557 ns/op
PASS
```

Heap-escape:
```
./keccakfast.go:17:6: cannot inline keccakReset: no function body
./keccakfast.go:21:6: cannot inline keccakWrite: no function body
./keccakfast.go:25:6: cannot inline keccakRead: no function body
./keccakfast.go:30:6: cannot inline hashPair: function too complex: cost 259 exceeds budget 80
./keccakfast.go:38:6: cannot inline hash: function too complex: cost 193 exceeds budget 80
./keccakfast.go:28:68: inlining call to sha3.NewLegacyKeccak256
./keccakfast.go:28:68: Before inlining: 
.   CALLFUNC hash.Hash tc(1) # keccakfast.go:28:68
./keccakfast.go:28:68: After inlining 
.   .   AS2 Def tc(1) # keccakfast.go:28:68
.   .   INLMARK # +keccakfast.go:28:68
.   INLCALL hash.Hash tc(1) # keccakfast.go:28:68
.   .   BLOCK tc(1) # keccakfast.go:28:68
.   .   .   DCL tc(1) # keccakfast.go:28:68
.   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
.   .   .   AS2 tc(1) # keccakfast.go:28:68
.   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
.   .   .   .   CONVIFACE Implicit hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   PTRLIT PTR-*sha3.state tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   STRUCTLIT sha3.state tc(1) # keccakfast.go:28:68 hashes.go:71:52
.   .   .   .   .   .   .   STRUCTKEY sha3.rate # keccakfast.go:28:68 hashes.go:71:53
.   .   .   .   .   .   .   .   LITERAL-136 int tc(1) # keccakfast.go:28:68 hashes.go:71:59
.   .   .   .   .   .   .   STRUCTKEY sha3.outputLen # keccakfast.go:28:68 hashes.go:71:64
.   .   .   .   .   .   .   .   LITERAL-32 int tc(1) # keccakfast.go:28:68 hashes.go:71:75
.   .   .   .   .   .   .   STRUCTKEY sha3.dsbyte # keccakfast.go:28:68 hashes.go:71:79
.   .   .   .   .   .   .   .   LITERAL-1 byte tc(1) # keccakfast.go:28:68 hashes.go:71:87
.   .   .   .   .   ADDR PTR-*uint8 tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   ADDR PTR-*uint8 tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   GOTO fast..i0 tc(1) # keccakfast.go:28:68
.   .   LABEL fast..i0 # keccakfast.go:28:68
.   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
./keccakfast.go:28:44: inlining call to reflect.ValueOf
./keccakfast.go:28:44: Before inlining: 
.   CALLFUNC reflect.Value tc(1) # keccakfast.go:28:44
.   .   CONVIFACE Implicit any tc(1) # keccakfast.go:28:68
.   .   .   PAREN hash.Hash tc(1) # keccakfast.go:28:68
.   .   .   .   .   AS2 Def tc(1) # keccakfast.go:28:68
.   .   .   .   .   INLMARK # +keccakfast.go:28:68
.   .   .   .   INLCALL hash.Hash tc(1) # keccakfast.go:28:68
.   .   .   .   .   BLOCK tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   DCL tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
.   .   .   .   .   .   AS2 tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
.   .   .   .   .   .   .   CONVIFACE Implicit hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   .   .   PTRLIT PTR-*sha3.state tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   .   .   .   STRUCTLIT sha3.state tc(1) # keccakfast.go:28:68 hashes.go:71:52
.   .   .   .   .   .   .   .   .   .   STRUCTKEY sha3.rate # keccakfast.go:28:68 hashes.go:71:53
.   .   .   .   .   .   .   .   .   .   .   LITERAL-136 int tc(1) # keccakfast.go:28:68 hashes.go:71:59
.   .   .   .   .   .   .   .   .   .   STRUCTKEY sha3.outputLen # keccakfast.go:28:68 hashes.go:71:64
.   .   .   .   .   .   .   .   .   .   .   LITERAL-32 int tc(1) # keccakfast.go:28:68 hashes.go:71:75
.   .   .   .   .   .   .   .   .   .   STRUCTKEY sha3.dsbyte # keccakfast.go:28:68 hashes.go:71:79
.   .   .   .   .   .   .   .   .   .   .   LITERAL-1 byte tc(1) # keccakfast.go:28:68 hashes.go:71:87
.   .   .   .   .   .   .   .   ADDR PTR-*uint8 tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   .   .   ADDR PTR-*uint8 tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   GOTO fast..i0 tc(1) # keccakfast.go:28:68
.   .   .   .   .   LABEL fast..i0 # keccakfast.go:28:68
.   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
./keccakfast.go:28:44: After inlining 
.   .   .   DCL # keccakfast.go:28:44
.   .   .   .   NAME-reflect.i Class:PAUTO Offset:0 InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3262:14
.   .   AS2 Def tc(1) # keccakfast.go:28:44
.   .   .   NAME-reflect.i Class:PAUTO Offset:0 InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3262:14
.   .   .   CONVIFACE Implicit any tc(1) # keccakfast.go:28:68
.   .   .   .   PAREN hash.Hash tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   AS2 Def tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   INLMARK # +keccakfast.go:28:68
.   .   .   .   .   INLCALL hash.Hash tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   BLOCK tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   .   DCL tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
.   .   .   .   .   .   .   AS2 tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
.   .   .   .   .   .   .   .   CONVIFACE Implicit hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   .   .   .   PTRLIT PTR-*sha3.state tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   .   .   .   .   STRUCTLIT sha3.state tc(1) # keccakfast.go:28:68 hashes.go:71:52
.   .   .   .   .   .   .   .   .   .   .   STRUCTKEY sha3.rate # keccakfast.go:28:68 hashes.go:71:53
.   .   .   .   .   .   .   .   .   .   .   .   LITERAL-136 int tc(1) # keccakfast.go:28:68 hashes.go:71:59
.   .   .   .   .   .   .   .   .   .   .   STRUCTKEY sha3.outputLen # keccakfast.go:28:68 hashes.go:71:64
.   .   .   .   .   .   .   .   .   .   .   .   LITERAL-32 int tc(1) # keccakfast.go:28:68 hashes.go:71:75
.   .   .   .   .   .   .   .   .   .   .   STRUCTKEY sha3.dsbyte # keccakfast.go:28:68 hashes.go:71:79
.   .   .   .   .   .   .   .   .   .   .   .   LITERAL-1 byte tc(1) # keccakfast.go:28:68 hashes.go:71:87
.   .   .   .   .   .   .   .   .   ADDR PTR-*uint8 tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   .   .   .   ADDR PTR-*uint8 tc(1) # keccakfast.go:28:68 hashes.go:71:46
.   .   .   .   .   .   .   GOTO fast..i0 tc(1) # keccakfast.go:28:68
.   .   .   .   .   .   LABEL fast..i0 # keccakfast.go:28:68
.   .   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack hash.Hash tc(1) # keccakfast.go:28:68 hashes.go:71:27
.   .   DCL # keccakfast.go:28:44
.   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3262:21
.   .   AS tc(1) # keccakfast.go:28:44
.   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3262:21
.   .   INLMARK Index:1 # +keccakfast.go:28:44
.   INLCALL reflect.Value tc(1) # keccakfast.go:28:44
.   .   IF tc(1) # keccakfast.go:28:44 value.go:3263:2
.   .   .   EQ bool tc(1) # keccakfast.go:28:44 value.go:3263:7
.   .   .   .   NAME-reflect.i Class:PAUTO Offset:0 InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3262:14
.   .   .   .   NIL any tc(1) # keccakfast.go:28:44 value.go:3263:10
.   .   .   BLOCK tc(1) # keccakfast.go:28:44
.   .   .   .   AS2 tc(1) # keccakfast.go:28:44
.   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3262:21
.   .   .   .   .   STRUCTLIT reflect.Value tc(1) # keccakfast.go:28:44 value.go:3264:15
.   .   .   .   GOTO fast..i1 tc(1) # keccakfast.go:28:44
.   .   BLOCK tc(1) # keccakfast.go:28:44
.   .   .   AS2 tc(1) # keccakfast.go:28:44
.   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3262:21
.   .   .   .   CALLFUNC reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   .   .   NAME-reflect.i Class:PAUTO Offset:0 InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3262:14
.   .   .   GOTO fast..i1 tc(1) # keccakfast.go:28:44
.   .   LABEL fast..i1 # keccakfast.go:28:44
.   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3262:21
./keccakfast.go:28:44: inlining call to reflect.unpackEface
./keccakfast.go:28:44: Before inlining: 
.   CALLFUNC reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   NAME-reflect.i Class:PAUTO Offset:0 InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3262:14
./keccakfast.go:28:44: After inlining 
.   .   .   DCL # keccakfast.go:28:44 value.go:3266:20
.   .   .   .   NAME-reflect.i Class:PAUTO Offset:0 Addrtaken InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:154:18
.   .   AS2 Def tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   NAME-reflect.i Class:PAUTO Offset:0 Addrtaken InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:154:18
.   .   .   NAME-reflect.i Class:PAUTO Offset:0 InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3262:14
.   .   DCL # keccakfast.go:28:44 value.go:3266:20
.   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:154:25
.   .   AS tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:154:25
.   .   INLMARK Index:2 # +keccakfast.go:28:44 value.go:3266:20
.   INLCALL reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   DCL # keccakfast.go:28:44 value.go:3266:20 value.go:155:2
.   .   .   .   NAME-reflect.e Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.EmptyInterface tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:155:2
.   .   AS Def tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:155:4
.   .   .   NAME-reflect.e Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.EmptyInterface tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:155:2
.   .   .   CONVNOP PTR-*abi.EmptyInterface tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:155:43
.   .   .   .   CONVNOP UNSAFEPTR-unsafe.Pointer tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:155:44
.   .   .   .   .   ADDR PTR-*any tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:155:44
.   .   .   .   .   .   NAME-reflect.i Class:PAUTO Offset:0 Addrtaken InlFormal OnStack Used any tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:154:18
.   .   .   DCL # keccakfast.go:28:44 value.go:3266:20 value.go:157:2
.   .   .   .   NAME-reflect.t Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:2
.   .   AS Def tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:4
.   .   .   NAME-reflect.t Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:2
.   .   .   DOTPTR fast.Type PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:8
.   .   .   .   NAME-reflect.e Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.EmptyInterface tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:155:2
.   .   IF tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:158:2
.   .   .   EQ bool tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:158:7
.   .   .   .   NAME-reflect.t Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:2
.   .   .   .   NIL PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:158:10
.   .   .   BLOCK tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   .   AS2 tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:154:25
.   .   .   .   .   STRUCTLIT reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:159:15
.   .   .   .   GOTO fast..i2 tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   DCL # keccakfast.go:28:44 value.go:3266:20 value.go:161:2
.   .   .   .   NAME-reflect.f Class:PAUTO Offset:0 InlLocal OnStack Used reflect.flag tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:161:2
.   .   AS Def tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:161:4
.   .   .   NAME-reflect.f Class:PAUTO Offset:0 InlLocal OnStack Used reflect.flag tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:161:2
.   .   .   CONV reflect.flag tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:161:18
.   .   .   .   CALLFUNC abi.Kind tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:161:18
.   .   .   .   .   METHEXPR fast.Kind FUNC-func(*abi.Type) abi.Kind tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:161:13
.   .   .   .   .   NAME-reflect.t Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:2
.   .   IF tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:162:2
.   .   .   CALLFUNC bool tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:162:17
.   .   .   .   METHEXPR fast.IfaceIndir FUNC-func(*abi.Type) bool tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:162:6
.   .   .   .   NAME-reflect.t Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:2
.   .   .   ASOP-OR AsOp:OR tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:163:5
.   .   .   .   NAME-reflect.f Class:PAUTO Offset:0 InlLocal OnStack Used reflect.flag tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:161:2
.   .   .   .   LITERAL-128 reflect.flag tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:163:8
.   .   BLOCK tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   AS2 tc(1) # keccakfast.go:28:44 value.go:3266:20
.   .   .   .   NAME-fast.~r0 Class:PAUTO Offset:0 InlFormal OnStack reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:154:25
.   .   .   .   STRUCTLIT reflect.Value tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:165:14
.   .   .   .   .   STRUCTKEY reflect.typ_ # keccakfast.go:28:44 value.go:3266:20 value.go:165:15
.   .   .   .   .   .   NAME-reflect.t Class:PAUTO Offset:0 InlLocal OnStack Used PTR-*abi.Type tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:157:2
.   .   .   .   .   STRUCTKEY reflect.ptr # keccakfast.go:28:44 value.go:3266:20 value.go:165:19
.   .   .   .   .   .   DOTPTR fast.Data UNSAFEPTR-unsafe.Pointer tc(1) # keccakfast.go:28:44 value.go:3266:20 value.go:165:19
…